Logout_Functionality|Logout "Back" Button userLogout.feature icerisindeki Logout Functionality'e ait Logout “Back” button Testin Scenario stepleri tamamlanmıştır. Test adımları LogoutStepDefs class'ında yer almaktadır. LoginStepDefs'te yer alan Logout işlemine yönelik stepler LogoutStepDefs'e taşınmıştır.
