+++ src/test/java/com/dgmarkt/pages/LoginPage.java
@@ -52,7 +52,7 @@ public class LoginPage extends BasePage {
     @FindBy(xpath = "//*[@id=\"pt-logout-link\"]")
     public WebElement myAccountLogoutButton;
 
-    @FindBy(xpath = "//*[@id=\"content\"]/div[1]/div/div/div/div/div/div[3]/div[4]/div/div/div/div[2]/div/button/span")
+    @FindBy(xpath = "//span[text()='Continue']")
     public WebElement popupContinueButton;
 
     @FindBy(xpath = "//*[@id=\"top-links\"]/ul/li/ul/li[1]/a")
