+++ src/test/resources/features/UserLogout.feature
@@ -1,4 +1,3 @@
-
 Feature: Logout Functionality
 
   Background:
@@ -11,7 +10,7 @@ Feature: Logout Functionality
     When The user clicks on My Account button and clicks on Login button after enters valid credentials
     Then The user verifies that welcome message is displayed
     When The user clicks on My Account button and clicks on Logout button
-    Then The user closes the Account Logout popup by clicking continue.
+    Then The user closes the Account Logout popup by clicking continue
 
 
   Scenario: Logout Negative Test
@@ -40,3 +39,12 @@ Feature: Logout Functionality
     Then The user verifies that welcome message is displayed
     And The user clicks on My Account button
     Then The user verifies that they are unable to log out by clicking the Downloads button from the opened dropdown
+
+
+    Scenario: Logout “Back” button
+    When The user clicks on My Account button and clicks on Login button after enters valid credentials
+    Then The user verifies that welcome message is displayed
+    When The user clicks on My Account button and clicks on Logout button
+    Then The user clicks on the browser “Back” button
+    Then The user closes the Sign Up Newsletter pop up
+    Then The user verifies that they are on the main page
